tests/approved/video/video_height_reflect.html

<!DOCTYPE html>
<html>
  <head>
    <title>Video Test: video_height_reflect</title>
    <link rel="author" title="Intel" href="http://www.intel.com" />
    <link rel="help" href="http://dev.w3.org/html5/spec/single-page.html#dom-dim-height" />
    <link rel="help" href="http://dev.w3.org/html5/spec/single-page.html#dom-dim-reflect" />
    <meta name="flags" content="" />
    <meta name="assert" content="Check that the height attribute reflects the height content attribute in the video element" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <video id="m">The user agent doesn't support media element.</video>
    <script type="text/javascript">
        var media = document.getElementById("m");
        var ULONG = {
            'DEFAULT'  : 0,
            'MIN'      : 0,
            'INTEGER'  : 480,
            'FLOAT'    : 480.5,
            'MAX'      : 2147483647,
            'UPPER'    : 2147483648,
            'NEGATIVE' : -480,
            'INVALID'  : 'INVALID'
        };
        
        test(function() {
            assert_false(media.height < ULONG.MIN, "media.height initial negative");
        }, "Check if video.height intial value is a valid non-negative integer in the range 0 to 2147483647 inclusive");
        
        [
            [ULONG.MIN,      ULONG.DEFAULT, "Check if video.height returns zero after the height attribute is set to zero"],
            [ULONG.MAX,      ULONG.MAX,     "Check if video.height returns maximum non-negative integer after the height attribute is set to the maximum non-negative integer"],
            [ULONG.INTEGER,  ULONG.INTEGER, "Check if video.height returns the integer number after the height attribute is set to a valid non-negative integer in the range 0 to 2147483647"],
            [ULONG.FLOAT,    ULONG.INTEGER, "Check if video.height returns a shortest integer number after the height attribute is set to a floating-point number"],
            [ULONG.UPPER,    ULONG.DEFAULT, "Check if video.height returns zero after the height attribute is set to a number greater than the maximum non-negative integer"],
            [ULONG.NEGATIVE, ULONG.DEFAULT, "Check if video.height returns zero after the height attribute is set to a negative integer"],
            [ULONG.INVALID,  ULONG.DEFAULT, "Check if video.height returns zero after the height attribute is set to an invlid non-negative integer"],
        ].forEach(function(attr) {
            test(function() {
                media.height = attr[0];
                assert_equals(media.height, attr[1], "media.height new value");
            }, attr[2]);
        });
    </script>
  </body>
</html>
